requirements.txt
# ── Web framework ──
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ── Message buses ──
boto3==1.34.131          # SQS + S3
pika==1.3.2              # RabbitMQ
confluent-kafka==2.4.0   # Kafka

# ── Document conversion ──
PyPDF2==3.0.1
pdfplumber==0.11.0
pdf2image==1.17.0        # PDF → images (uses poppler)
pytesseract==0.3.10      # OCR
Pillow==10.3.0
python-docx==1.1.2       # DOCX
openpyxl==3.1.4          # XLSX
python-pptx==0.6.23      # PPTX
beautifulsoup4==4.12.3   # HTML
lxml==5.2.2
odfpy==1.4.1             # ODF / ODT
striprtf==0.0.26         # RTF
python-magic==0.4.27     # MIME detection
chardet==5.2.0

# ── File fetchers ──
requests==2.32.3         # HTTP/URL fetch
ftplib-enhanced==0.0.0;python_version<"0"   # stdlib ftplib is used directly

# ── Utilities ──
pydantic==2.7.4
pydantic-settings==2.3.4
